Also show status for merges from integr. branches into release branches
These merges are equally important as the merges from one release
branch into the next integration branches and they can fail as well.

In contrast to the merges from a release branch into an integration
branch they don't fail because of merge conflicts, but because the
code doesn't compile anymore or a test fails in the integration branch.