Add box for deployments of internal test installations
The installations are updated twice a day (if there are changes in
the corresponding branches) and the deployments should complete
without problems. If there is a problem, it's something we have to
fix (either it's really a bug or it's a flaky test or DB change set
or something like that that has to be fixed).